project-05-task-tracker-with-OAuth/api-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
PyJWT>=2.9.0
requests>=2.31.0
